% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_data.R
\name{gen_data}
\alias{gen_data}
\title{Generate data sets based on the provided parameters}
\usage{
gen_data(sims, n_subj, tau_1, beta_1, n_cycles)
}
\arguments{
\item{sims}{number of simulations to generate}

\item{n_subj}{number of subjects in one data set}

\item{tau_1}{by-subject random slope sd}

\item{beta_1}{expected change in tau from one time point to the next}

\item{n_cycles}{number of time points for each subject}
}
\value{
data.table of randomly generated data sets
}
\description{
Generate data sets based on the provided parameters
}
