% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{glmmPQL_model}
\alias{glmmPQL_model}
\title{Run glmmPQL on simulated data}
\usage{
glmmPQL_model(df)
}
\arguments{
\item{df}{data.table}
}
\value{
results of glmmPQL
}
\description{
Run glmmPQL on simulated data
}
