% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{glmmPQL_modelcorAR}
\alias{glmmPQL_modelcorAR}
\title{Run glmmPQL on simulated data using correlation}
\usage{
glmmPQL_modelcorAR(df)
}
\arguments{
\item{df}{dataframe of simulated data}
}
\value{
results from glmmPQL
}
\description{
Run glmmPQL on simulated data using correlation
}
