% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{lme_model_corAR}
\alias{lme_model_corAR}
\title{Run poslme on simulated data}
\usage{
lme_model_corAR(df)
}
\arguments{
\item{df}{dataframe of simulated data}
}
\value{
results from lme
}
\description{
Run poslme on simulated data
}
