% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{lmer_model}
\alias{lmer_model}
\title{run lmer on simulated data}
\usage{
lmer_model(df)
}
\arguments{
\item{df}{dataframe of simulated data}
}
\value{
results from lmer
}
\description{
run lmer on simulated data
}
